doc: fix ci, move wavetank to components
